+++ railties/lib/rails/dev_caching.rb
@@ -15,6 +15,7 @@ module Rails
         end
 
         FileUtils.touch 'tmp/restart.txt'
+        FileUtils.rm_f('tmp/pids/server.pid')
       end
 
       def enable_by_argument(caching)